DataProcessing/requirements.txt
azure-functions==1.0.0a4
azure-functions-worker==1.0.0a4
grpcio==1.14.2
grpcio-tools==1.14.2
numpy==1.15.4
pandas==0.23.4
protobuf==3.6.1
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
